pyproject.toml
[tool.poetry]
name = "doorlockd-client"
version = "0.1.0"
description = ""
authors = ["Diederik Hamelink"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pyaml = "^23.9.7"
simplejson = "^3.19.2"
toml = "^0.10.2"
nfcpy = {git = "https://github.com/matthijskooijman/nfcpy.git", rev = "v1.0.4+doorlockd", optional = true}
requests = "^2.31.0"
gpiod = {version = "^2.1.3" , optional = true}

[tool.poetry.extras]
gpiod = ["gpiod"]
nfcpy = ["nfcpy"]

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
